+++ modules/integration/emr-to-opensearch/stack.py
@@ -174,12 +174,10 @@ class EMRtoOpensearch(Stack):  # type: ignore
                 {
                     "id": "AwsSolutions-IAM4",
                     "reason": "Managed Policies are for service account roles only",
-                    "applies_to": "*",
                 },
                 {
                     "id": "AwsSolutions-IAM5",
                     "reason": "Resource access restriced to ADDF resources",
-                    "applies_to": "*",
                 },
             ],
         )