+++ modules/post-processing/yolop-lane-detection/src/detect_lanes.py
@@ -89,7 +89,9 @@ def detect(cfg, opt):
     if not os.path.exists(opt.json_path):
         os.mkdir(opt.json_path)
 
-    for i, (path, img, img_det, vid_cap, shapes) in tqdm(enumerate(dataset), total=len(dataset)):
+    for i, (path, img, img_det, vid_cap, shapes) in tqdm(
+        enumerate(dataset), total=len(dataset)
+    ):
         img = transform(img).to(device)
         img = img.half() if half else img.float()  # uint8 to fp16/32
         if img.ndimension() == 3:
@@ -106,7 +108,11 @@ def detect(cfg, opt):
         # Apply NMS
         t3 = time_synchronized()
         det_pred = non_max_suppression(
-            inf_out, conf_thres=opt.conf_thres, iou_thres=opt.iou_thres, classes=None, agnostic=False
+            inf_out,
+            conf_thres=opt.conf_thres,
+            iou_thres=opt.iou_thres,
+            classes=None,
+            agnostic=False,
         )
         t4 = time_synchronized()
 
@@ -127,26 +133,38 @@ def detect(cfg, opt):
         ratio = shapes[1][0][1]
 
         da_predict = da_seg_out[:, :, pad_h : (height - pad_h), pad_w : (width - pad_w)]
-        da_seg_mask = torch.nn.functional.interpolate(da_predict, scale_factor=int(1 / ratio), mode="bilinear")
+        da_seg_mask = torch.nn.functional.interpolate(
+            da_predict, scale_factor=int(1 / ratio), mode="bilinear"
+        )
         _, da_seg_mask = torch.max(da_seg_mask, 1)
         da_seg_mask = da_seg_mask.int().squeeze().cpu().numpy()
         # da_seg_mask = morphological_process(da_seg_mask, kernel_size=7)
 
         ll_predict = ll_seg_out[:, :, pad_h : (height - pad_h), pad_w : (width - pad_w)]
-        ll_seg_mask = torch.nn.functional.interpolate(ll_predict, scale_factor=int(1 / ratio), mode="bilinear")
+        ll_seg_mask = torch.nn.functional.interpolate(
+            ll_predict, scale_factor=int(1 / ratio), mode="bilinear"
+        )
         _, ll_seg_mask = torch.max(ll_seg_mask, 1)
         ll_seg_mask = ll_seg_mask.int().squeeze().cpu().numpy()
         # Lane line post-processing
         # ll_seg_mask = morphological_process(ll_seg_mask, kernel_size=7, func_type=cv2.MORPH_OPEN)
         # ll_seg_mask = connect_lane(ll_seg_mask)
 
-        img_det = show_seg_result(img_det, (da_seg_mask, ll_seg_mask), _, _, is_demo=True)
+        img_det = show_seg_result(
+            img_det, (da_seg_mask, ll_seg_mask), _, _, is_demo=True
+        )
 
         if len(det):
             det[:, :4] = scale_coords(img.shape[2:], det[:, :4], img_det.shape).round()
             for *xyxy, conf, cls in reversed(det):
                 label_det_pred = f"{names[int(cls)]} {conf:.2f}"
-                plot_one_box(xyxy, img_det, label=label_det_pred, color=colors[int(cls)], line_thickness=2)
+                plot_one_box(
+                    xyxy,
+                    img_det,
+                    label=label_det_pred,
+                    color=colors[int(cls)],
+                    line_thickness=2,
+                )
 
         # if dataset.mode == 'images':
         cv2.imwrite(save_path, img_det)
@@ -177,21 +195,48 @@ def detect(cfg, opt):
 
 if __name__ == "__main__":
     parser = argparse.ArgumentParser()
-    parser.add_argument("--weights", nargs="+", type=str, default="weights/End-to-end.pth", help="model.pth path(s)")
+    parser.add_argument(
+        "--weights",
+        nargs="+",
+        type=str,
+        default="weights/End-to-end.pth",
+        help="model.pth path(s)",
+    )
     parser.add_argument(
         "--source", type=str, default="/opt/ml/processing/input/image", help="source"
     )  # file/folder   ex:inference/images
-    parser.add_argument("--img-size", type=int, default=640, help="inference size (pixels)")
-    parser.add_argument("--conf-thres", type=float, default=0.25, help="object confidence threshold")
-    parser.add_argument("--iou-thres", type=float, default=0.45, help="IOU threshold for NMS")
-    parser.add_argument("--device", default="cpu", help="cuda device, i.e. 0 or 0,1,2,3 or cpu")
     parser.add_argument(
-        "--save_dir", type=str, default="/opt/ml/processing/output/image", help="directory to save results"
+        "--img-size", type=int, default=640, help="inference size (pixels)"
+    )
+    parser.add_argument(
+        "--conf-thres", type=float, default=0.25, help="object confidence threshold"
+    )
+    parser.add_argument(
+        "--iou-thres", type=float, default=0.45, help="IOU threshold for NMS"
+    )
+    parser.add_argument(
+        "--device", default="cpu", help="cuda device, i.e. 0 or 0,1,2,3 or cpu"
+    )
+    parser.add_argument(
+        "--save_dir",
+        type=str,
+        default="/opt/ml/processing/output/image",
+        help="directory to save results",
     )
     parser.add_argument("--augment", action="store_true", help="augmented inference")
     parser.add_argument("--update", action="store_true", help="update all models")
-    parser.add_argument("--csv_path", type=str, default="/opt/ml/processing/output/csv", help="output path for csv")
-    parser.add_argument("--json_path", type=str, default="/opt/ml/processing/output/json", help="output path for json")
+    parser.add_argument(
+        "--csv_path",
+        type=str,
+        default="/opt/ml/processing/output/csv",
+        help="output path for csv",
+    )
+    parser.add_argument(
+        "--json_path",
+        type=str,
+        default="/opt/ml/processing/output/json",
+        help="output path for json",
+    )
     opt = parser.parse_args()
     with torch.no_grad():
         detect(cfg, opt)