+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - added lane detection module for building docker image into ECR
 - object and lane detection added to the rosbag-image-pipeline
 - added lane detection WITH YOLOP module for building docker image into ECR
+- replace lanedet with yolop for lane detection pipeline
 
 ### **Changed**